unify tests
